+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 6.0.2
 -----
 
-Release date: ``2021-0x-xx``
+Release date: ``2021-06-07``
 
 - `NXPY-225 <https://jira.nuxeo.com/browse/NXPY-225>`__: Allow to pass additional arguments when asking for a ``batchId``