+++ CHANGELOG.rst
@@ -7,11 +7,12 @@ Changelog
 Release date: ``2021-0x-xx``
 
 - `NXPY-228 <https://jira.nuxeo.com/browse/NXPY-228>`__: Use proper job names in GitHub workflows
+- `NXPY-230 <https://jira.nuxeo.com/browse/NXPY-230>`__: Allow to pass requests arguments to the OAuth2 client
 
 Technical changes
 -----------------
 
--
+- Added ``subclient_kwargs`` keyword argument to ``OAuth2.__init__()``
 
 6.0.2
 -----